+++ libleptongui/include/find_text_state.h
@@ -1,7 +1,7 @@
 /* Lepton EDA Schematic Capture
  * Copyright (C) 1998-2010 Ales Hvezda
  * Copyright (C) 1998-2015 gEDA Contributors
- * Copyright (C) 2017-2024 Lepton EDA Contributors
+ * Copyright (C) 2017-2026 Lepton EDA Contributors
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -56,6 +56,7 @@ struct _SchematicFindTextState
   SchematicBin parent;
 
   GtkListStore *store;
+  GtkWidget *tree_widget;
 };
 
 
@@ -67,7 +68,7 @@ G_BEGIN_DECLS
 int
 schematic_find_text_state_find (SchematicWindow *w_current,
                                 SchematicFindTextState *state,
-                                GList *pages,
+                                GSList *all_pages,
                                 int type,
                                 const char *text,
                                 gboolean descend,
@@ -75,6 +76,12 @@ schematic_find_text_state_find (SchematicWindow *w_current,
 GtkWidget*
 schematic_find_text_state_new ();
 
+GtkTreeSelection*
+schematic_find_text_state_get_selection (SchematicFindTextState *state);
+
+void
+schematic_find_text_state_select (GtkTreeSelection *selection,
+                                  SchematicFindTextState *state);
 G_END_DECLS
 
 #endif /* _FIND_TEXT_STATE_H_ */